requirements.txt
# Twarga Cloud MVP - Requirements

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.36
alembic==1.13.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.6

# Templates and Static Files
jinja2==3.1.2
python-multipart==0.0.6

# System Monitoring
psutil==5.9.6

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# CORS is included in FastAPI, no separate package needed

# Environment Variables
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# JSON Schema (using Python 3.13 compatible versions)
pydantic==2.10.5
pydantic-settings==2.6.1

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: Chart.js integration (if needed)
# plotly==5.17.0

# Optional: Advanced logging
# structlog==23.2.0

# VM Management (Vagrant integration)
# vagrant==1.0.1  # Note: This might need custom implementation

# Terminal Access (ttyd integration)
# websockets==12.0